create extra files for ble, wifi and telnet functions
